UNTESTED: Allow commands to respond to both IRC and Discord.
GenChannels represent replying to both an IRC channel and a Discord
channel (as well as making messages generic in that same sense).  Allows
Sintia to reply to both IRC and Discord, and also paves the path for
multi-channel support by beginning to pass IRC channels around.

Parses successfully, but otherwise is untested.